[Refactor] 풀리퀘 건의사항 반영 -#53
